pyproject.toml
[project]
name = "portfolio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.46.0",
    "monsterui>=1.0.29",
    "numpy>=2.3.4",
    "python-fasthtml>=0.12.29",
]
